fix: shadow lambda parameters and resolve mapping-receiver aliases; split scope module
_locally_bound_names()/_enclosing_qualnames() deliberately don't model a
lambda as its own scope -- a lambda's body shares its enclosing
function's qualname, so a lambda's own parameter was never recorded as
bound anywhere _shadowed()'s qualname-based check could see. Fixed by
having _shadowed() also walk the call's true AST ancestry, checking
every enclosing ast.Lambda's own parameters directly -- exact by
construction, so it can never misattribute a shadow to a call genuinely
outside the lambda even when it shares the same qualname.

_is_mapping_receiver() only matched vars(rec)/X.__dict__ directly --
fields = vars(rec); fields["bases"] was invisible once the mapping was
stored in an intermediate variable first. Fixed with a new
_mapping_receiver_aliases(), the identical name-only fixed-point pattern
every other alias helper in this module already uses.

These two fixes pushed the file to 1975/2000 lines. Split
scripts/fact_field_readers_scope.py out (mechanical extraction, mirroring
fact_detector_misuse_scope.py's identical split): _enclosing_qualnames,
_parent_map, _TRANSPARENT_EXPR_WRAPPER_TYPES, _outermost_containing_expr,
_locally_bound_names, _lexical_function_parents -- moved unchanged, with
a matching sys.path guard verified by running the wrapper-scoping test
file in isolation. Registered in scripts/CLAUDE.md's Inventory table.
